+++ src/test/java/service/FileServiceTest.java
@@ -1,0 +1,4 @@
+package service;
+
+class FileServiceTest {
+}
